# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API PU-SDA-JATIM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "resetPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "password_confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/import/file/parent": {
            "post": {
                "tags": [
                    "Import"
                ],
                "operationId": "importfileparent",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "description": "file",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/import/file/children/{id}": {
            "post": {
                "tags": [
                    "Import"
                ],
                "operationId": "importfilechildren",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "description": "file",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}